+++ R/flywire-coord.R
@@ -3,7 +3,7 @@
 #' @importFrom jsonlite toJSON
 map1 <- function(xyz1, scale=2) {
   xyz1=as.integer(round(xyz1))
-  baseurl <- "https://tbar.uvm.edu/app/flyconv/dataset/flywire_v1/"
+  baseurl <- "https://spine.janelia.org/app/flyconv/dataset/flywire_v1/"
   url <- sprintf(paste0(baseurl, "s/%d/z/%d/x/%d/y/%d/"),
                  scale, xyz1[3], xyz1[1], xyz1[2])
   res = GET(url)
@@ -23,30 +23,53 @@ map1 <- function(xyz1, scale=2) {
   unlist(res)
 }
 
-mapmany <- function(xyz, scale=2, ...) {
+
+#' @importFrom httr content_type
+mapmany <- function(xyz, scale=2, msgpack=NULL, round=TRUE, ...) {
   if(!is.matrix(xyz) || ncol(xyz)!=3)
     stop("I need an Nx3 matrix of points!")
   xyz=round(xyz)
-  baseurl <- "https://tbar.uvm.edu/app/flyconv/dataset/flywire_v1"
-  url <- sprintf("%s/s/%d/values", baseurl, scale)
-  body <- list(locations=xyz)
-  bodyj <- toJSON(body, auto_unbox=FALSE)
-  res = POST(url, body = bodyj, config = content_type_json(), encode='raw', ...)
-  if(status_code(res)>400) {
-    warn_for_status(res)
+  # because we should be rounding to nearest voxel I think
+  # maybe check this with Eric Perlman. Should def be case for z.
+  if(round)
+    mode(xyz)='integer'
+  baseurl <- "https://spine.janelia.org/app/flyconv/dataset/flywire_v1"
+  url <- sprintf("%s/s/%d/values_array", baseurl, scale)
+  if(is.null(msgpack))
+    msgpack <- requireNamespace('RcppMsgPack', quietly = TRUE)
+  body <- list(x=xyz[,1], y=xyz[,2], z=xyz[,3])
+  # msgpack doesn't handle length 1 arrays
+  msgpack=msgpack && length(body$x)>1
+  resp <- if(msgpack) {
+    bodym=RcppMsgPack::msgpack_pack(body)
+    POST(url, body = bodym, config = content_type("application/msgpack"),
+         encode='raw', ...)
+  } else {
+    bodyj <- toJSON(body, auto_unbox=FALSE)
+    POST(url, body = bodyj, config = content_type_json(), encode='raw', ...)
+  }
+  if(status_code(resp)>400) {
+    warn_for_status(resp)
     badval=matrix(NA_real_, ncol = 5, nrow=nrow(xyz))
     colnames(badval)=c("dx", "dy", "x", "y", "z")
     return(badval)
   }
-
-  strres = content(
-    res,
-    as = 'text',
-    type = 'application/json',
-    encoding = 'utf-8',
-  )
-  strres=gsub("NaN", '"NA"', strres, fixed = TRUE)
-  jsonlite::fromJSON(strres, simplifyVector = TRUE)
+  res <- if(msgpack) {
+    rawres=content(resp, as='raw',type = 'application/msgpack')
+    RcppMsgPack::msgpack_unpack(rawres, simplify = T)
+  } else {
+    strres = content(
+      resp,
+      as = 'text',
+      type = 'application/json',
+      encoding = 'utf-8',
+    )
+    strres=gsub("NaN", '"NA"', strres, fixed = TRUE)
+    jsonlite::fromJSON(strres, simplifyVector = TRUE)
+  }
+  cols=do.call(cbind, res)
+  colnames(cols)=c("dx", "dy", "x", "y", "z")
+  cols
 }
 
 #' Map points in  FlyWire v1 TO FAFB14 space (xyz nm)
@@ -107,7 +130,7 @@ mapmany <- function(xyz, scale=2, ...) {
 #' AV4b1.flywire <- xform_brain(AV4b1, reference="FlyWire", sample="FAFB14")
 #' plot3d(neuronlist(AV4b1.flywire, AV4b1))
 #' }
-flywire2fafb <- function(xyz, method=c("mapmany", "map1"), chunksize=200,
+flywire2fafb <- function(xyz, method=c("mapmany", "map1"), chunksize=40e3,
                          swap=FALSE, ...) {
   if(!isTRUE(length(dim(xyz))==2))
     stop("Please give me N x 3 points as input!")